refactor: split out Mayan route types and classes
